+++ src/components/beta/gux-chart-line/gux-chart-line.tsx
@@ -30,15 +30,19 @@ export class GuxLineChart {
       point: false
     },
     config: {
+      axis: {
+        ticks: false,
+        titlePadding: 8
+      },
+      axisX: {
+        labelAngle: 0
+      },
       legend: {
         symbolType: 'circle'
       }
     },
     encoding: {
-      x: {
-        type: 'nominal',
-        axis: { labelAngle: 0 } // horizontal x axis ticks by default
-      },
+      x: { type: 'nominal' },
       y: { type: 'quantitative' },
       color: {
         field: DEFAULT_COLOR_FIELD_NAME,
@@ -66,6 +70,18 @@ export class GuxLineChart {
 
   @Prop()
   includeLegend: boolean;
+
+  @Prop()
+  legendPosition:
+    | 'left'
+    | 'right'
+    | 'top'
+    | 'bottom'
+    | 'top-left'
+    | 'top-right'
+    | 'bottom-left'
+    | 'bottom-right'
+    | 'none' = 'right';
 
   @Prop()
   includeDataPointMarkers: boolean;
@@ -129,11 +145,15 @@ export class GuxLineChart {
     }
 
     if (this.xTickLabelSlant) {
-      this.baseChartSpec.encoding.x.axis.labelAngle = 45;
+      this.baseChartSpec.config.axisX.labelAngle = 45;
     }
 
     if (this.includeLegend) {
       this.baseChartSpec.encoding.color.legend = true;
+    }
+
+    if (this.legendPosition) {
+      this.baseChartSpec.config.legend.orient = this.legendPosition;
     }
 
     const xFieldName = this.xFieldName;